pyproject.toml
[project]
name = "provas"
version = "0.1.0"
description = ""
authors = [
    {name = "jlxmns",email = "jlxmns@pm.me"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "django-ninja (>=1.4.0,<2.0.0)",
    "django-ninja-jwt (>=5.3.7,<6.0.0)"
]

[tool.ruff]
fix = true
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]
exclude = ["**/migrations"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
ruff = "^0.11.4"

